+++ client/src/pages/Configurations/ConfigTest/index.test.tsx
@@ -7,7 +7,7 @@ import userEvent, { UserEvent } from '@testing-library/user-event';
 import { useRunInlineConfigurationTest } from '../../../api/configurations/configurations';
 import { Mock } from 'vitest';
 import { AxiosError } from 'axios';
-import { mockCustomCodes, mockCodeSets } from '../ManageCodes/fixtures';
+import { mockCustomCodes, mockCodeSets } from '../test/fixtures';
 
 // Mock all API requests.
 
@@ -52,6 +52,7 @@ vi.mock('../../../api/configurations/configurations', async () => {
             { id: 'gonorrhea-1', display_name: 'Gonorrhea', associated: false },
           ],
           all_versions: [{ version: 1 }],
+          rsg_codes: [],
         },
       },
     })),
@@ -94,7 +95,6 @@ describe('Config testing page', () => {
 
   it('should show "Test configuration" as the current step', () => {
     renderPage();
-    expect(screen.getByText('Build', { selector: 'a' })).toBeInTheDocument();
     expect(screen.getByText('Test', { selector: 'a' })).toHaveAttribute(
       'aria-current',
       'page'
